+++ app/models/developer.rb
@@ -53,7 +53,7 @@ class Developer < ApplicationRecord
   scope :newest_first, -> { order(created_at: :desc) }
   scope :visible, -> { where.not(search_status: :invisible).or(where(search_status: nil)) }
 
-  after_create_commit :send_admin_notification
+  after_create_commit :send_admin_notification, :send_welcome_email
 
   def visible?
     !invisible?
@@ -89,6 +89,10 @@ class Developer < ApplicationRecord
 
   def send_admin_notification
     NewDeveloperProfileNotification.with(developer: self).deliver_later(User.admin)
+  end
+
+  def send_welcome_email
+    DeveloperMailer.with(developer: self).welcome_email.deliver_later
   end
 
   def send_invisiblize_notification