+++ test/mailers/previews/developer_mailer_preview.rb
@@ -3,4 +3,8 @@ class DeveloperMailerPreview < ActionMailer::Preview
     notification = Notification.where(type: InvisiblizeDeveloperNotification.to_s).first
     DeveloperMailer.with(record: notification, recipient: notification.recipient).invisiblize
   end
+
+  def welcome_developer
+    DeveloperMailer.with(developer: Developer.first).welcome_email
+  end
 end