+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### April
 
+* April 18 - Automatically send Developer and Business welcome emails - #378
 * April 18 - Added banner to connect with Joe to pay hiring fee - #380
 * April 12 - Added messaging tips for businesses - #376
 * April 10 - Added "Invisiblize" button on developer profiles (for admins) - #370